SpringTeleEcom Version 2: Fix Lombok processor config, update JWT Service API, add @Builder to models

+++ src/main/java/com/example/SpringTeleEcom/model/dto/OrderItemRequest.java
@@ -1,6 +1,6 @@
 package com.example.SpringTeleEcom.model.dto;
 
 public record OrderItemRequest(
-    int productId,
+    long productId,
     int quantity
 ){}
